rename openingHoursToOpeningHoursSpecification and put it with other json-ld util functions
