+++ common/utils/json-ld.ts
@@ -2,6 +2,11 @@ import { convertImageUri } from './convert-image-uri';
 import { Organization } from '../model/organization';
 import { BreadcrumbItems } from '../model/breadcrumbs';
 import { JsonLdObj } from '@weco/common/views/components/JsonLd/JsonLd';
+import type {
+  OpeningHours,
+  OpeningHoursDay,
+  SpecialOpeningHours,
+} from '../model/opening-hours';
 
 export function objToJsonLd<T>(obj: T, type: string, root = true) {
   const jsonObj = JSON.parse(JSON.stringify(obj));
@@ -80,4 +85,38 @@ function imageLd(image: Image) {
       'ImageObject'
     )
   );
+}
+
+export function openingHoursLd(openingHours: OpeningHours | undefined): {
+  openingHoursSpecification: OpeningHoursDay[];
+  specialOpeningHoursSpecification: SpecialOpeningHours[];
+} {
+  return {
+    openingHoursSpecification: openingHours?.regular
+      ? openingHours.regular.map(openingHoursDay => {
+          const specObject = objToJsonLd(
+            {
+              dayOfWeek: openingHoursDay.dayOfWeek,
+              opens: openingHoursDay.opens,
+              closes: openingHoursDay.closes,
+            },
+            'OpeningHoursSpecification',
+            false
+          );
+          delete specObject.note;
+          return specObject;
+        })
+      : [],
+    specialOpeningHoursSpecification: openingHours?.exceptional
+      ? openingHours.exceptional.map(openingHoursDate => {
+          const specObject = {
+            opens: openingHoursDate.opens,
+            closes: openingHoursDate.closes,
+            validFrom: openingHoursDate.overrideDate?.format('DD MMMM YYYY'),
+            validThrough: openingHoursDate.overrideDate?.format('DD MMMM YYYY'),
+          };
+          return objToJsonLd(specObject, 'OpeningHoursSpecification', false);
+        })
+      : [],
+  };
 }
